Consolidate auxiliary functions in file_store.utilities
